+++ src/Pim/Bundle/DataGridBundle/Datagrid/Product/ContextConfigurator.php
@@ -90,25 +90,19 @@ class ContextConfigurator implements ConfiguratorInterface
     protected $gridViewRepository;
 
     /**
-     * @param DatagridConfiguration    $configuration
      * @param ProductManager           $productManager
      * @param RequestParameters        $requestParams
-     * @param Request                  $request
      * @param SecurityContextInterface $securityContext
      * @param EntityRepository         $gridViewRepository
      */
     public function __construct(
-        DatagridConfiguration $configuration,
         ProductManager $productManager,
         RequestParameters $requestParams,
-        Request $request,
         SecurityContextInterface $securityContext,
         EntityRepository $gridViewRepository
     ) {
-        $this->configuration      = $configuration;
         $this->productManager     = $productManager;
         $this->requestParams      = $requestParams;
-        $this->request            = $request;
         $this->securityContext    = $securityContext;
         $this->gridViewRepository = $gridViewRepository;
     }
@@ -116,8 +110,9 @@ class ContextConfigurator implements ConfiguratorInterface
     /**
      * {@inheritdoc}
      */
-    public function configure()
+    public function configure(DatagridConfiguration $configuration)
     {
+        $this->configuration = $configuration;
         $this->addProductStorage();
         $this->addLocaleCode();
         $this->addScopeCode();
@@ -128,6 +123,14 @@ class ContextConfigurator implements ConfiguratorInterface
         $this->addDisplayedColumnCodes();
         $this->addAttributesIds();
         $this->addAttributesConfig();
+    }
+
+    /**
+     * @param Request $request
+     */
+    public function setRequest(Request $request = null)
+    {
+        $this->request = $request;
     }
 
     /**
@@ -146,15 +149,25 @@ class ContextConfigurator implements ConfiguratorInterface
     protected function addAttributesIds()
     {
         $attributeCodes = $this->getUserGridColumns();
-        $repository     = $this->productManager->getAttributeRepository();
-        $attributeIds   = ($attributeCodes) ? $repository->getAttributeIds($attributeCodes) : null;
-
-        if (!$attributeIds) {
-            $attributeIds = $repository->getAttributeIdsUseableInGrid();
-        }
+        $attributeIds = $this->getAttributeIds($attributeCodes);
 
         $path = $this->getSourcePath(self::DISPLAYED_ATTRIBUTES_KEY);
         $this->configuration->offsetSetByPath($path, $attributeIds);
+    }
+
+    /**
+     * Return useable attribute ids
+     *
+     * @param string[] $attributeCodes
+     *
+     * @return integer[]
+     */
+    protected function getAttributeIds($attributeCodes = null)
+    {
+        $repository   = $this->productManager->getAttributeRepository();
+        $attributeIds = $repository->getAttributeIdsUseableInGrid($attributeCodes);
+
+        return $attributeIds;
     }
 
     /**
@@ -319,9 +332,13 @@ class ContextConfigurator implements ConfiguratorInterface
      */
     protected function getAttributesConfig()
     {
-        $repository     = $this->productManager->getAttributeRepository();
-        $attributeIds  = $repository->getAttributeIdsUseableInGrid();
+        $attributeIds  = $this->getAttributeIds();
+        if (empty($attributeIds)) {
+            return [];
+        }
+
         $currentLocale = $this->getCurrentLocaleCode();
+        $repository    = $this->productManager->getAttributeRepository();
         $configuration = $repository->getAttributesAsArray(true, $currentLocale, $attributeIds);
 
         return $configuration;
@@ -335,7 +352,6 @@ class ContextConfigurator implements ConfiguratorInterface
     protected function getUserGridColumns()
     {
         $params = $this->requestParams->get(RequestParameters::ADDITIONAL_PARAMETERS);
-
         if (isset($params['view']) && isset($params['view']['columns'])) {
             return explode(',', $params['view']['columns']);
         }